+++ app/src/main/res/layout/activity_signuppage.xml
@@ -8,7 +8,7 @@
     android:orientation="vertical"
     android:gravity="center"
     android:background="@color/white"
-    tools:context=".SignuppageActivity">
+    tools:context=".account.SignuppageActivity">
 
     <TextView
         android:layout_width="wrap_content"
@@ -83,6 +83,7 @@
     </com.google.android.material.textfield.TextInputLayout>
 
     <android.widget.Button
+        android:id="@+id/make_account"
         android:layout_width="300dp"
         android:layout_height="wrap_content"
         android:layout_marginTop="25dp"
