+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,7 @@
-<resources>
+resources>
     <string name="app_name">Mycrodiary</string>
+    <string name="title_activity_mypage">Mypage</string>
+    <string name="title_home">Home</string>
+    <string name="title_dashboard">Dashboard</string>
+    <string name="title_notifications">Notifications</string>
 </resources>
