+++ app/src/main/res/values/themes.xml
@@ -3,6 +3,7 @@
     <style name="Base.Theme.Mycrodiary" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
+        <item name="colorPrimary">@color/main_color</item>
     </style>
 
     <style name="Theme.Mycrodiary" parent="Base.Theme.Mycrodiary" />
